backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "firebase-admin>=7.1.0",
    "google-genai>=1.60.0",
    "load-dotenv>=0.1.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
